restructuring to use src folder
