+++ src/lib/browser.ts
