+++ src/styles/animated-gradient.css
